moving around the new templatefunctions
